requirements.txt
bleak==0.22.3
influxdb==5.3.2
python-dotenv==1.0.1
